+++ .gitignore
@@ -29,7 +29,6 @@ Temporary Items
 /node_modules
 /dist
 /target
-/lib
 
 # -----------------------------------
 # Files
